+++ src/routes/playlist/[id]/+page.svelte
@@ -657,10 +657,40 @@
 		<div {@attach virtual_grid.attach()}>
 			<div class="drag-line" class:hidden={drag_to_index === null} bind:this={drag_line}></div>
 		</div>
+		{#if tracks_page.itemIds.length === 0}
+			<div class="filter-guide absolute inset-0 flex items-center justify-center">
+				<div class="w-full max-w-lg px-6 py-10 text-center">
+					<div class="text-base font-medium">Filter syntax</div>
+
+					<div class="mt-5 space-y-2 text-left text-sm">
+						<div class="rounded-md bg-gray-300/5 px-3 py-2.5">
+							<span>"the one"</span>
+						</div>
+						<div class="rounded-md bg-gray-300/5 px-3 py-2.5">
+							<span>artist:apashe</span>
+						</div>
+						<div class="rounded-md bg-gray-300/5 px-3 py-2.5">
+							<span>album:"7 minutes dead"</span>
+						</div>
+					</div>
+					<div class="mt-4 text-sm text-white/50">
+						Results include words with extra punctuation and accents.
+						<br />
+						So <span>FOOL</span> matches <span>F.O.O.L</span>, but <span>F.O.Ó.L</span> does not
+						match
+						<span>FOOL</span>.
+					</div>
+				</div>
+			</div>
+		{/if}
 	</div>
 </div>
 
 <style lang="sass">
+	.filter-guide span
+		background-color: hsla(var(--hue), 20%, 42%, 0.3)
+		padding: 2px 4px
+		border-radius: 3px
 	.tracklist :global
 		.odd
 			background-color: hsla(0, 0%, 90%, 0.06)