+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 ## Next
 - Remove macOS 10.15 support
 - Add exact filtering with double quotes, like `"the one"`
+- Show filter syntax guide on empty page
 - Add syntax highlighting for filters like `album:"mirai sekai"`
 - Remove the active filter highlight
 - Fix undo/redo in the filter textbox only changing one character at a time